requirements.txt
pbr>=0.6,!=0.7,<1.0
oslo.config>=2.3.0
oslo.log>=1.8.0
oslo.utils

six >=1.4.0

python-novaclient
python-keystoneclient
python-glanceclient
python-ceilometerclient


dirq
python-dateutil
